♻️ refactor(eval): move loading UI into feature boundary
