+++ src/features/Eval/index.ts
@@ -1,0 +1,1 @@
+export { default as BenchmarkDetailLoading } from './BenchmarkDetailLoading';
